+++ Assets/Scripts/TES/NIF/NIFObjectBuilder.cs
@@ -228,13 +228,14 @@ namespace TESUnity
 					meshRenderer.enabled = false;
 				}
 
-                obj.isStatic = true;
+                if (TESManager.instance.useStaticBatching)
+                    obj.isStatic = true;
 			}
 			
 			if(collidable)
 			{
 				obj.AddComponent<MeshCollider>().sharedMesh = mesh;
-				if(TESUnity.instance.useKinematicRigidbodies)
+				if(TESManager.instance.useKinematicRigidbodies)
 				{
 					obj.AddComponent<Rigidbody>().isKinematic = true;
 				}