+++ CHANGES.md
@@ -1,15 +1,29 @@
 #TESUnity Changelog
 
+<<<<<<< HEAD
 ## Version 0.7 - 03/xx/2018
 ### Added
 - Lightweight Scriptable Render Pipeline Support
 
 ### Changed
 - Force Vulkan on Linux as first API (OpenGL as Fallback)
+=======
+## Version 0.7 - 03/27/2018
+### Added
+- Lightweight Render Pipeline for better performances
+- New exposed variables for more performances
+- Use Vulkan as first GFX API on Linux
+
+### Changed
+- Updated to Unity 2018.1b12
+- Improved performances
+
+### Fixed
+- Specular value is set to 0.25 instead of 0.5
+>>>>>>> xr
 
 ## Version 0.6 - 03/14/2018
 ### Added
-- Updated to Unity 2018.1b10
 - New VR integration
 - Controller support
 - Auto Normal Map generation
@@ -17,6 +31,9 @@
 - Normal Map support
 - Day/Night Cycle
 - Graphics Improvement
+
+### Changed
+- Updated to Unity 2018.1b10
 
 ### Removed
 - OSVR Support